bbhm_patch.sh: revert back to setting SYSCFG_DB_FILE to /nvram/syscfg.db in all cases
The recent upstream change to control SYSCFG_DB_FILE based on
BOX_TYPE is not easy to maintain. Hardcode the value to match the
legacy behaviour when the UpdateNvram syscfg value was set to true.

It's not necessary (or safe - since the value may not exist) to test
the UpdateNvram syscfg value.

Note also that this bbhm_patch.sh script is not installed on the
target, so this is just generic cleanup...

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>