drop unused duplicate copies of psm_hal_apis.c from source-arm and source-pc
The version of psm_hal_apis.c in source-arm and source-pc are not
used during the build process. Since they are the same as the default
version in source/Ssp/psm_hal_apis.c there's no real value in keeping
them as a reference. Drop them as a cleanup and to avoid confusion.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>