RDKBDEV-2564: improved error checking for calls to..
Reason for change:

improved error checking for calls to AnscXmlDomNodeGetDataString()

Check for strings returned without a nul terminator by comparing the
length of the string returned with the size of the target buffer.

Should be merged with or after the corresponding cleanup in
ccsp-common-library.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

minor code cleanup in loadRecordFromXML()

Although length will be ignored (since target buffer pointer is
NULL), it's not clean or safe to rely on that.

Also drop unnecessary call to AnscXmlDomNodeGetDataBinary() in error
case.

Test Procedure: Sanity.
Risks: None.

Change-Id: I2d233693fc54deedd7acad8e1b03e8925163b24d
Signed-off-by: Andre McCurdy <armccurdy@gmail.com>